+++ src/main/java/io/github/xwasu/LangRepository.java
@@ -9,11 +9,11 @@ class LangRepository {
 
     LangRepository() {
         languages = new ArrayList<>();
-        languages.add(new Lang(1L, "Hello", "en"));
-        languages.add(new Lang(2L, "Siemanko", "pl"));
+        languages.add(new Lang(1, "Hello", "en"));
+        languages.add(new Lang(2, "Siemanko", "pl"));
     }
 
-    Optional<Lang> findById(Long id) {
+    Optional<Lang> findById(Integer id) {
         return languages.stream()
                 .filter(l -> l.getId().equals(id))
                 .findFirst();
